back/Contact/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.1
python-dotenv==1.0.0
pyjwt==2.8.0
pydantic==2.5.0
email-validator==2.0.0